+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceEnrollmentPlatformRestriction/MSFT_IntuneDeviceEnrollmentPlatformRestriction.psm1
@@ -54,6 +54,18 @@ function Get-TargetResource
         $MacOSRestriction,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $PlatformRestriction,
+
+        [Parameter()]
+        [ValidateSet('android', 'androidForWork', 'ios', 'mac', 'windows')]
+        $PlatformType,
+
+        [Parameter()]
+        [System.Int32]
+        $Priority,
+
+        [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $Assignments,
 
@@ -107,51 +119,184 @@ function Get-TargetResource
 
     try
     {
-        $config = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -DeviceEnrollmentConfigurationId $Identity -ErrorAction silentlyContinue
+        $getValue = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -DeviceEnrollmentConfigurationId $Identity -ErrorAction SilentlyContinue
 
-        if ($null -eq $config)
+        if ($null -eq $getValue)
         {
-            Write-Verbose -Message "No Device Enrollment Platform Restriction {$Identity} was found. Trying to retrieve instance by name {$DisplayName}"
-            $config = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -Filter "DisplayName eq '$DisplayName'" `
-                -ErrorAction silentlyContinue
-            if ($null -eq $config)
-            {
-                Write-Verbose -Message "No instances found by name {$DisplayName}"
-                return $nullResult
-            }
+            Write-Verbose -Message "Could not find an Intune Device Enrollment Platform Restriction with Id {$Identity}"
+
+            $getValue = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -Filter "DisplayName eq '$DisplayName'" `
+                -ErrorAction SilentlyContinue | Where-Object `
+                -FilterScript { `
+                    $_.AdditionalProperties.'@odata.type' -like "#microsoft.graph.deviceEnrollmentPlatformRestriction*Configuration" -and `
+                    $(if ($null -ne $_.AdditionalProperties.platformType) { $_.AdditionalProperties.platformType -eq $PlatformType } else { $true }) `
+                }
         }
 
-        Write-Verbose -Message "Found Device Enrollment Platform Restriction with Name {$($config.DisplayName)}"
+        if ($null -eq $getValue)
+        {
+            Write-Verbose -Message "Could not find an Intune Device Enrollment Platform Restriction with DisplayName {$DisplayName}"
+            return $nullResult
+        }
+
+        Write-Verbose -Message "Found Intune Device Enrollment Platform Restriction with Name {$($getValue.DisplayName)}"
         $results = @{
-            Identity                          = $config.Id
-            DisplayName                       = $config.DisplayName
-            Description                       = $config.Description
-            DeviceEnrollmentConfigurationType = $config.DeviceEnrollmentConfigurationType.toString()
+            Identity                          = $getValue.Id
+            DisplayName                       = $getValue.DisplayName
+            Description                       = $getValue.Description
+            DeviceEnrollmentConfigurationType = $getValue.DeviceEnrollmentConfigurationType.ToString()
+            Priority                          = $getValue.Priority
             Ensure                            = 'Present'
             Credential                        = $Credential
             ApplicationId                     = $ApplicationId
             TenantId                          = $TenantId
             ApplicationSecret                 = $ApplicationSecret
             CertificateThumbprint             = $CertificateThumbprint
-            Managedidentity                   = $ManagedIdentity.IsPresent
+            ManagedIdentity                   = $ManagedIdentity.IsPresent
         }
 
-        $results += Get-DevicePlatformRestrictionSetting -Properties $config.AdditionalProperties
+        # Check if it is not a "Default platform restriction"
+        if ($getValue.AdditionalProperties.platformType)
+        {
+            $results.Add('PlatformType', $getValue.AdditionalProperties.platformType.ToString())
+
+            $complexPlatformRestriction = @{}
+            $complexPlatformRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.platformRestriction.blockedManufacturers)
+            $complexPlatformRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.platformRestriction.blockedSkus)
+            $complexPlatformRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.platformRestriction.osMaximumVersion)
+            $complexPlatformRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.platformRestriction.osMinimumVersion)
+            $complexPlatformRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.platformRestriction.personalDeviceEnrollmentBlocked)
+            $complexPlatformRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.platformRestriction.platformBlocked)
+            if ($complexPlatformRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexPlatformRestriction = $null
+            }
+
+            $results.Add("PlatformRestriction", $complexPlatformRestriction)
+        } 
+        else
+        {
+            $complexAndroidForWorkRestriction = @{}
+            $complexAndroidForWorkRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.androidForWorkRestriction.blockedManufacturers)
+            $complexAndroidForWorkRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.androidForWorkRestriction.blockedSkus)
+            $complexAndroidForWorkRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.androidForWorkRestriction.osMaximumVersion)
+            $complexAndroidForWorkRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.androidForWorkRestriction.osMinimumVersion)
+            $complexAndroidForWorkRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.androidForWorkRestriction.personalDeviceEnrollmentBlocked)
+            $complexAndroidForWorkRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.androidForWorkRestriction.platformBlocked)
+            if ($complexAndroidForWorkRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexAndroidForWorkRestriction = $null
+            }
+
+            $complexAndroidRestriction = @{}
+            $complexAndroidRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.androidRestriction.blockedManufacturers)
+            $complexAndroidRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.androidRestriction.blockedSkus)
+            $complexAndroidRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.androidRestriction.osMaximumVersion)
+            $complexAndroidRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.androidRestriction.osMinimumVersion)
+            $complexAndroidRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.androidRestriction.personalDeviceEnrollmentBlocked)
+            $complexAndroidRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.androidRestriction.platformBlocked)
+            if ($complexAndroidRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexAndroidRestriction = $null
+            }
+
+            $complexIosRestriction = @{}
+            $complexIosRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.iosRestriction.blockedManufacturers)
+            $complexIosRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.iosRestriction.blockedSkus)
+            $complexIosRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.iosRestriction.osMaximumVersion)
+            $complexIosRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.iosRestriction.osMinimumVersion)
+            $complexIosRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.iosRestriction.personalDeviceEnrollmentBlocked)
+            $complexIosRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.iosRestriction.platformBlocked)
+            if ($complexIosRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexIosRestriction = $null
+            }
+
+            $complexMacOSRestriction = @{}
+            $complexMacOSRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.macOSRestriction.blockedManufacturers)
+            $complexMacOSRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.macOSRestriction.blockedSkus)
+            $complexMacOSRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.macOSRestriction.osMaximumVersion)
+            $complexMacOSRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.macOSRestriction.osMinimumVersion)
+            $complexMacOSRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.macOSRestriction.personalDeviceEnrollmentBlocked)
+            $complexMacOSRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.macOSRestriction.platformBlocked)
+            if ($complexMacOSRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexMacOSRestriction = $null
+            }
+
+            $complexMacRestriction = @{}
+            $complexMacRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.macRestriction.blockedManufacturers)
+            $complexMacRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.macRestriction.blockedSkus)
+            $complexMacRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.macRestriction.osMaximumVersion)
+            $complexMacRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.macRestriction.osMinimumVersion)
+            $complexMacRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.macRestriction.personalDeviceEnrollmentBlocked)
+            $complexMacRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.macRestriction.platformBlocked)
+            if ($complexMacRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexMacRestriction = $null
+            }
+
+            $complexWindowsHomeSkuRestriction = @{}
+            $complexWindowsHomeSkuRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.windowsHomeSkuRestriction.blockedManufacturers)
+            $complexWindowsHomeSkuRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.windowsHomeSkuRestriction.blockedSkus)
+            $complexWindowsHomeSkuRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.windowsHomeSkuRestriction.osMaximumVersion)
+            $complexWindowsHomeSkuRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.windowsHomeSkuRestriction.osMinimumVersion)
+            $complexWindowsHomeSkuRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.windowsHomeSkuRestriction.personalDeviceEnrollmentBlocked)
+            $complexWindowsHomeSkuRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.windowsHomeSkuRestriction.platformBlocked)
+            if ($complexWindowsHomeSkuRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexWindowsHomeSkuRestriction = $null
+            }
+
+            $complexWindowsMobileRestriction = @{}
+            $complexWindowsMobileRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.windowsMobileRestriction.blockedManufacturers)
+            $complexWindowsMobileRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.windowsMobileRestriction.blockedSkus)
+            $complexWindowsMobileRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.windowsMobileRestriction.osMaximumVersion)
+            $complexWindowsMobileRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.windowsMobileRestriction.osMinimumVersion)
+            $complexWindowsMobileRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.windowsMobileRestriction.personalDeviceEnrollmentBlocked)
+            $complexWindowsMobileRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.windowsMobileRestriction.platformBlocked)
+            if ($complexWindowsMobileRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexWindowsMobileRestriction = $null
+            }
+
+            $complexWindowsRestriction = @{}
+            $complexWindowsRestriction.Add('BlockedManufacturers', $getValue.AdditionalProperties.windowsRestriction.blockedManufacturers)
+            $complexWindowsRestriction.Add('BlockedSkus', $getValue.AdditionalProperties.windowsRestriction.blockedSkus)
+            $complexWindowsRestriction.Add('OsMaximumVersion', $getValue.AdditionalProperties.windowsRestriction.osMaximumVersion)
+            $complexWindowsRestriction.Add('OsMinimumVersion', $getValue.AdditionalProperties.windowsRestriction.osMinimumVersion)
+            $complexWindowsRestriction.Add('PersonalDeviceEnrollmentBlocked', $getValue.AdditionalProperties.windowsRestriction.personalDeviceEnrollmentBlocked)
+            $complexWindowsRestriction.Add('PlatformBlocked', $getValue.AdditionalProperties.windowsRestriction.platformBlocked)
+            if ($complexWindowsRestriction.values.Where({$null -ne $_}).count -eq 0)
+            {
+                $complexWindowsRestriction = $null
+            }
+
+            $results.Add("AndroidForWorkRestriction", $complexAndroidForWorkRestriction)
+            $results.Add("AndroidRestriction", $complexAndroidRestriction)
+            $results.Add("IosRestriction", $complexIosRestriction)
+            $results.Add("MacOSRestriction", $complexMacOSRestriction)
+            $results.Add("MacRestriction", $complexMacRestriction)
+            $results.Add("WindowsHomeSkuRestriction", $complexWindowsHomeSkuRestriction)
+            $results.Add("WindowsMobileRestriction", $complexWindowsMobileRestriction)
+            $results.Add("WindowsRestriction", $complexWindowsRestriction)
+        }
 
         if ($null -ne $results.WindowsMobileRestriction)
         {
             $results.Remove('WindowsMobileRestriction') | Out-Null
         }
 
-        $AssignmentsValues = Get-MgBetaDeviceManagementDeviceEnrollmentConfigurationAssignment -DeviceEnrollmentConfigurationId $config.Id
+        $assignmentsValues = Get-MgBetaDeviceManagementDeviceEnrollmentConfigurationAssignment -DeviceEnrollmentConfigurationId $getValue.Id
         $assignmentResult = @()
-        foreach ($assignmentEntry in $AssignmentsValues)
+        foreach ($assignmentEntry in $assignmentsValues)
         {
             $assignmentValue = @{
-                dataType                                   = $assignmentEntry.Target.AdditionalProperties.'@odata.type'
-                deviceAndAppManagementAssignmentFilterType = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType.ToString()
-                deviceAndAppManagementAssignmentFilterId   = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterId
-                groupId                                    = $assignmentEntry.Target.AdditionalProperties.groupId
+                dataType = $assignmentEntry.Target.AdditionalProperties.'@odata.type'
+                deviceAndAppManagementAssignmentFilterType = $(if ($null -ne $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType)
+                    {$assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterType.ToString()})
+                deviceAndAppManagementAssignmentFilterId = $assignmentEntry.Target.DeviceAndAppManagementAssignmentFilterId
+                groupId = $assignmentEntry.Target.AdditionalProperties.groupId
             }
             $assignmentResult += $assignmentValue
         }
@@ -226,6 +371,18 @@ function Set-TargetResource
         $MacOSRestriction,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $PlatformRestriction,
+
+        [Parameter()]
+        [ValidateSet('android', 'androidForWork', 'ios', 'mac', 'windows')]
+        $PlatformType,
+
+        [Parameter()]
+        [System.Int32]
+        $Priority,
+
+        [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $Assignments,
 
@@ -273,78 +430,90 @@ function Set-TargetResource
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
 
-    $currentCategory = Get-TargetResource @PSBoundParameters
-    $PSBoundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
-    $PSBoundParameters.Remove('Identity') | Out-Null
+    if (-not [System.String]::IsNullOrEmpty($PlatformType) -and $null -eq $PlatformRestriction) {
+        throw 'If PlatformType is specified, PlatformRestriction is required.'
+    }
 
-    if ($Ensure -eq 'Present' -and $currentCategory.Ensure -eq 'Absent')
+    if ([System.String]::IsNullOrEmpty($PlatformType) -and $null -ne $PlatformRestriction) {
+        throw 'PlatformRestriction can only be set on policies with a PlatformType.'
+    }
+
+    if ($Ensure -eq 'Absent' -and $Identity -like '*_DefaultPlatformRestrictions') {
+        throw 'Cannot delete the default platform restriction policy.'
+    }
+
+    $currentInstance = Get-TargetResource @PSBoundParameters
+
+    $BoundParameters = Remove-M365DSCAuthenticationParameter -BoundParameters $PSBoundParameters
+
+    $BoundParameters.Remove('Identity') | Out-Null
+
+    if ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Absent')
     {
-        Write-Verbose -Message "Creating new Device Enrollment Platform Restriction {$DisplayName}"
+        Write-Verbose -Message "Creating an Intune Device Enrollment Platform Restriction with DisplayName {$DisplayName}"
 
-        $PSBoundParameters.Remove('Assignments') | Out-Null
+        $BoundParameters.Remove('Assignments') | Out-Null
 
-        if ($PSBoundParameters.Keys.Contains('WindowsMobileRestriction'))
+        $CreateParameters = ([Hashtable]$BoundParameters).Clone()
+        $CreateParameters = Rename-M365DSCCimInstanceParameter -Properties $CreateParameters
+        $CreateParameters.Remove('Id') | Out-Null
+
+        if ($BoundParameters.Keys.Contains('WindowsMobileRestriction'))
         {
             if ($WindowsMobileRestriction.platformBlocked -eq $false)
             {
                 Write-Verbose -Message 'Windows Mobile platform is deprecated and cannot be unblocked, reverting back to blocked'
-
                 $WindowsMobileRestriction.platformBlocked = $true
             }
         }
 
-        $keys = (([Hashtable]$PSBoundParameters).clone()).Keys
+        $keys = (([Hashtable]$CreateParameters).Clone()).Keys
         foreach ($key in $keys)
         {
-            $keyName = $key.substring(0, 1).toLower() + $key.substring(1, $key.length - 1)
-            $keyValue = $PSBoundParameters.$key
-            if ($null -ne $PSBoundParameters.$key -and $PSBoundParameters.$key.getType().Name -like '*cimInstance*')
+            if ($null -ne $CreateParameters.$key -and $CreateParameters.$key.GetType().Name -like '*cimInstance*')
             {
-                $keyValue = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $PSBoundParameters.$key
-                if ($DeviceEnrollmentConfigurationType -eq 'singlePlatformRestriction' )
-                {
-                    $keyName = 'platformRestriction'
-                    $PSBoundParameters.add('platformType', ($key.replace('Restriction', '')))
-                }
+                $CreateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $CreateParameters.$key
             }
-            $PSBoundParameters.remove($key)
-            $PSBoundParameters.add($keyName, $keyValue)
         }
 
-        $policyType = '#microsoft.graph.deviceEnrollmentPlatformRestrictionConfiguration'
-        if ($DeviceEnrollmentConfigurationType -eq 'platformRestrictions' )
+        # Check if it is a "Default platform restriction"
+        if ([System.String]::IsNullOrEmpty($PlatformType))
         {
-            $policyType = '#microsoft.graph.deviceEnrollmentPlatformRestrictionsConfiguration'
-            $PSBoundParameters.add('deviceEnrollmentConfigurationType ', 'limit')
+            $CreateParameters.Add('@odata.type', '#microsoft.graph.deviceEnrollmentPlatformRestrictionsConfiguration')
         }
-        $PSBoundParameters.add('@odata.type', $policyType)
+        else
+        {
+            $CreateParameters.Add('@odata.type', '#microsoft.graph.deviceEnrollmentPlatformRestrictionConfiguration')
+        }
+        
+        $policy = New-MgBetaDeviceManagementDeviceEnrollmentConfiguration -BodyParameter $CreateParameters
 
-        #Write-Verbose ($PSBoundParameters | ConvertTo-Json -Depth 20)
+        $assignmentsHash = @()
+        foreach ($assignment in $Assignments)
+        {
+            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+        }
 
-        $policy = New-MgBetaDeviceManagementDeviceEnrollmentConfiguration `
-            -BodyParameter ([hashtable]$PSBoundParameters)
-
-        #Assignments from DefaultPolicy are not editable and will raise an alert
+        # Skip for the default platform restriction
         if ($policy.Id -notlike '*_DefaultPlatformRestrictions')
         {
-            if ($null -ne $Assignments -and $Assignments -ne @())
-            {
-                $assignmentsHash = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignments
+            Update-DeviceConfigurationPolicyAssignment -DeviceConfigurationPolicyId  $policy.Id `
+                -Targets $assignmentsHash `
+                -Repository 'deviceManagement/deviceEnrollmentConfigurations' `
+                -RootIdentifier 'enrollmentConfigurationAssignments'
+        }
 
-                Update-DeviceConfigurationPolicyAssignment `
-                    -DeviceConfigurationPolicyId  $policy.id `
-                    -Targets $assignmentsHash `
-                    -Repository 'deviceManagement/deviceEnrollmentConfigurations'
-            }
+        if ($policy.Priority -ne $Priority)
+        {
+            Write-Warning -Message 'Priority of the new policy is not equal to the specified priority. To solve this issue, reapply the configuration or make sure that the lowest priority policies are applied after the highest priority ones.'
         }
     }
-    elseif ($Ensure -eq 'Present' -and $currentCategory.Ensure -eq 'Present')
+    elseif ($Ensure -eq 'Present' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Updating Device Enrollment Platform Restriction {$DisplayName}"
+        Write-Verbose -Message "Updating the Intune Device Enrollment Platform Restriction with Id {$($currentInstance.Identity)}"
+        $BoundParameters.Remove('Assignments') | Out-Null
 
-        $PSBoundParameters.Remove('Assignments') | Out-Null
-
-        if ($PSBoundParameters.Keys.Contains('WindowsMobileRestriction'))
+        if ($BoundParameters.Keys.Contains('WindowsMobileRestriction'))
         {
             if ($WindowsMobileRestriction.platformBlocked -eq $false)
             {
@@ -354,55 +523,63 @@ function Set-TargetResource
             }
         }
 
-        $keys = (([Hashtable]$PSBoundParameters).clone()).Keys
+        $UpdateParameters = ([Hashtable]$BoundParameters).clone()
+        $UpdateParameters = Rename-M365DSCCimInstanceParameter -Properties $UpdateParameters
+
+        $UpdateParameters.Remove('Id') | Out-Null
+        $UpdateParameters.Remove('Priority') | Out-Null
+
+        $keys = (([Hashtable]$UpdateParameters).clone()).Keys
         foreach ($key in $keys)
         {
-            $keyName = $key.substring(0, 1).toLower() + $key.substring(1, $key.length - 1)
-            $keyValue = $PSBoundParameters.$key
-            if ($null -ne $PSBoundParameters.$key -and $PSBoundParameters.$key.getType().Name -like '*cimInstance*')
+            if ($null -ne $UpdateParameters.$key -and $UpdateParameters.$key.getType().Name -like '*cimInstance*')
             {
-                $keyValue = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $PSBoundParameters.$key
-                if ($DeviceEnrollmentConfigurationType -eq 'singlePlatformRestriction' )
-                {
-                    $keyName = 'platformRestriction'
-                }
+                $UpdateParameters.$key = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $UpdateParameters.$key
             }
-            $PSBoundParameters.remove($key)
-            $PSBoundParameters.add($keyName, $keyValue)
         }
 
-        $policyType = '#microsoft.graph.deviceEnrollmentPlatformRestrictionConfiguration'
-        if ($DeviceEnrollmentConfigurationType -eq 'platformRestrictions' )
+        # Check if it is a "Default platform restriction"
+        if ($currentInstance.Identity -like "*_DefaultPlatformRestrictions")
         {
-            $policyType = '#microsoft.graph.deviceEnrollmentPlatformRestrictionsConfiguration'
+            $UpdateParameters.Add("@odata.type", "#microsoft.graph.deviceEnrollmentPlatformRestrictionsConfiguration")
         }
-        $PSBoundParameters.add('@odata.type', $policyType)
-        #Write-Verbose ($PSBoundParameters | ConvertTo-Json -Depth 20)
-        Update-MgBetaDeviceManagementDeviceEnrollmentConfiguration `
-            -BodyParameter ([hashtable]$PSBoundParameters) `
-            -DeviceEnrollmentConfigurationId $Identity
-
-        #Assignments from DefaultPolicy are not editable and will raise an alert
-        if ($Identity -notlike '*_DefaultPlatformRestrictions')
+        else
         {
-            if ($null -ne $Assignments -and $Assignments -ne @())
+            $UpdateParameters.Add("@odata.type", "#microsoft.graph.deviceEnrollmentPlatformRestrictionConfiguration")
+            $UpdateParameters.Remove("PlatformType")
+        }
+
+        Update-MgBetaDeviceManagementDeviceEnrollmentConfiguration  `
+            -DeviceEnrollmentConfigurationId $currentInstance.Identity `
+            -BodyParameter $UpdateParameters
+
+        $assignmentsHash = @()
+        foreach ($assignment in $Assignments)
+        {
+            $assignmentsHash += Get-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignment
+        }
+
+        # Skip for the default platform restriction
+        if ($currentInstance.Identity -notlike "*_DefaultPlatformRestrictions")
+        {
+            if ($Priority -ne $currentInstance.Priority)
             {
-                $assignmentsHash = Convert-M365DSCDRGComplexTypeToHashtable -ComplexObject $Assignments
-
-                Update-DeviceConfigurationPolicyAssignment `
-                -DeviceConfigurationPolicyId  $Identity `
+                $uri = "/beta/deviceManagement/deviceEnrollmentConfigurations/$($currentInstance.Identity)/setpriority"
+                Invoke-MgGraphRequest -Method POST -Uri $uri -Body $(@{ 'priority' = $Priority} | ConvertTo-Json) -ErrorAction Stop
+            }
+            
+            Update-DeviceConfigurationPolicyAssignment `
+                -DeviceConfigurationPolicyId $currentInstance.Identity `
                 -Targets $assignmentsHash `
-                -Repository 'deviceManagement/deviceEnrollmentConfigurations'
-            }
+                -Repository 'deviceManagement/deviceEnrollmentConfigurations' `
+                -RootIdentifier 'enrollmentConfigurationAssignments'
         }
     }
-    elseif ($Ensure -eq 'Absent' -and $currentCategory.Ensure -eq 'Present')
+    elseif ($Ensure -eq 'Absent' -and $currentInstance.Ensure -eq 'Present')
     {
-        Write-Verbose -Message "Removing Device Enrollment Platform Restriction {$DisplayName}"
-        $config = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -Filter "displayName eq '$DisplayName'" `
-        | Where-Object -FilterScript { $_.AdditionalProperties.'@odata.type' -eq '#microsoft.graph.deviceEnrollmentPlatformRestrictionsConfiguration' }
+        Write-Verbose -Message "Removing the Intune Device Enrollment Platform Restriction with Id {$($currentInstance.Identity)}"
 
-        Remove-MgBetaDeviceManagementDeviceEnrollmentConfiguration -DeviceEnrollmentConfigurationId $config.id
+        Remove-MgBetaDeviceManagementDeviceEnrollmentConfiguration -DeviceEnrollmentConfigurationId $currentInstance.Identity
     }
 }
 
@@ -462,6 +639,18 @@ function Test-TargetResource
         $MacOSRestriction,
 
         [Parameter()]
+        [Microsoft.Management.Infrastructure.CimInstance]
+        $PlatformRestriction,
+
+        [Parameter()]
+        [ValidateSet('android', 'androidForWork', 'ios', 'mac', 'windows')]
+        $PlatformType,
+
+        [Parameter()]
+        [System.Int32]
+        $Priority,
+
+        [Parameter()]
         [Microsoft.Management.Infrastructure.CimInstance[]]
         $Assignments,
 
@@ -505,10 +694,10 @@ function Test-TargetResource
         -Parameters $PSBoundParameters
     Add-M365DSCTelemetryEvent -Data $data
     #endregion
-    Write-Verbose -Message "Testing configuration of Device Enrollment Platform Restriction {$DisplayName}"
+    Write-Verbose -Message "Testing configuration of the Intune Device Enrollment Platform Restriction with Id {$Identity} and DisplayName {$DisplayName}"
 
     $CurrentValues = Get-TargetResource @PSBoundParameters
-    $ValuesToCheck = ([Hashtable]$PSBoundParameters).clone()
+    $ValuesToCheck = ([Hashtable]$PSBoundParameters).Clone()
 
     if ($CurrentValues.Ensure -ne $Ensure)
     {
@@ -544,29 +733,18 @@ function Test-TargetResource
     $ValuesToCheck.Remove('Identity') | Out-Null
     $ValuesToCheck.Remove('WindowsMobileRestriction') | Out-Null
 
-    #Convert any DateTime to String
-    foreach ($key in $ValuesToCheck.Keys)
-    {
-        if (($null -ne $CurrentValues[$key]) `
-                -and ($CurrentValues[$key].getType().Name -eq 'DateTime'))
-        {
-            $CurrentValues[$key] = $CurrentValues[$key].toString()
-        }
-    }
-
     Write-Verbose -Message "Current Values: $(Convert-M365DscHashtableToString -Hashtable $CurrentValues)"
     Write-Verbose -Message "Target Values: $(Convert-M365DscHashtableToString -Hashtable $PSBoundParameters)"
     #Compare basic parameters
     if ($testResult)
     {
-        Write-Verbose -Message "Comparing the current values with the desired ones"
         $testResult = Test-M365DSCParameterState -CurrentValues $CurrentValues `
             -Source $($MyInvocation.MyCommand.Source) `
             -DesiredValues $PSBoundParameters `
             -ValuesToCheck $ValuesToCheck.Keys
     }
 
-    Write-Verbose -Message "Test-TargetResource returned $TestResult"
+    Write-Verbose -Message "Test-TargetResource returned $testResult"
 
     return $TestResult
 }
@@ -577,10 +755,6 @@ function Export-TargetResource
     [OutputType([System.String])]
     param
     (
-        [Parameter()]
-        [System.String]
-        $Filter,
-
         [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
@@ -622,8 +796,10 @@ function Export-TargetResource
 
     try
     {
-        [array]$configs = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration -All:$true -Filter $Filter -ErrorAction Stop `
-        | Where-Object -FilterScript { $_.AdditionalProperties.'@odata.type' -like '#microsoft.graph.deviceEnrollmentPlatform*Configuration' }
+        [array]$configs = Get-MgBetaDeviceManagementDeviceEnrollmentConfiguration `
+            -All `
+            -Filter "deviceEnrollmentConfigurationType eq 'singlePlatformRestriction'" `
+            -ErrorAction Stop
 
         $i = 1
         $dscContent = ''
@@ -647,9 +823,12 @@ function Export-TargetResource
                 TenantId              = $TenantId
                 ApplicationSecret     = $ApplicationSecret
                 CertificateThumbprint = $CertificateThumbprint
-                Managedidentity       = $ManagedIdentity.IsPresent
+                ManagedIdentity       = $ManagedIdentity.IsPresent
             }
+            
             $Results = Get-TargetResource @Params
+            $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
+                -Results $Results
 
             if ($null -ne $Results.Assignments)
             {
@@ -767,24 +946,28 @@ function Export-TargetResource
                 }
             }
 
+            if ($null -ne $Results.PlatformRestriction)
+            {
+                $complexTypeStringResult = Get-M365DSCDRGComplexTypeToString -ComplexObject ($Results.PlatformRestriction) -CIMInstanceName DeviceEnrollmentPlatformRestriction
+                if ($complexTypeStringResult)
+                {
+                    $Results.PlatformRestriction = $complexTypeStringResult
+                }
+                else
+                {
+                    $Results.Remove('PlatformRestriction') | Out-Null
+                }
+            }
 
-            $Results = Update-M365DSCExportAuthenticationResults -ConnectionMode $ConnectionMode `
-                -Results $Results
             $currentDSCBlock = Get-M365DSCExportContentForResource -ResourceName $ResourceName `
                 -ConnectionMode $ConnectionMode `
                 -ModulePath $PSScriptRoot `
                 -Results $Results `
                 -Credential $Credential
 
-
             if ($null -ne $Results.Assignments)
             {
-                $isCIMArray = $false
-                if ($Results.Assignments.getType().Fullname -like '*[[\]]')
-                {
-                    $isCIMArray = $true
-                }
-                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'Assignments' -IsCIMArray:$isCIMArray
+                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'Assignments' -IsCIMArray:$true
             }
 
             if ($null -ne $Results.IosRestriction)
@@ -801,6 +984,7 @@ function Export-TargetResource
             {
                 $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'WindowsHomeSkuRestriction'
             }
+
             if ($null -ne $Results.WindowsMobileRestriction)
             {
                 $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'WindowsMobileRestriction'
@@ -824,6 +1008,11 @@ function Export-TargetResource
             if ($null -ne $Results.MacOSRestriction)
             {
                 $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'MacOSRestriction'
+            }
+
+            if ($null -ne $Results.PlatformRestriction)
+            {
+                $currentDSCBlock = Convert-DSCStringParamToVariable -DSCBlock $currentDSCBlock -ParameterName 'PlatformRestriction'
             }
 
             $dscContent += $currentDSCBlock
@@ -854,98 +1043,6 @@ function Export-TargetResource
 
         return ''
     }
-}
-
-function Get-DevicePlatformRestrictionSetting
-{
-    [CmdletBinding()]
-    [OutputType([System.Collections.Hashtable])]
-    param
-    (
-        [Parameter(Mandatory = 'true')]
-        [System.Collections.Hashtable]
-        $Properties
-    )
-
-    $results = @{}
-
-    if ($null -ne $Properties.platformType)
-    {
-        $keyName = ($Properties.platformType).Substring(0, 1).toUpper() + ($Properties.platformType).substring(1, $Properties.platformType.length - 1) + 'Restriction'
-        $keyValue = [Hashtable]::new($Properties.platformRestriction)
-        $hash = @{}
-        foreach ($key in $keyValue.Keys)
-        {
-            if ($null -ne $keyValue.$key)
-            {
-                switch -Wildcard ($keyValue.$key.getType().name)
-                {
-                    '*[[\]]'
-                    {
-                        if ($keyValue.$key.count -gt 0)
-                        {
-                            $hash.add($key, $keyValue.$key)
-                        }
-                    }
-                    'String'
-                    {
-                        if (-Not [String]::IsNullOrEmpty($keyValue.$key))
-                        {
-                            $hash.add($key, $keyValue.$key)
-                        }
-                    }
-                    Default
-                    {
-                        $hash.add($key, $keyValue.$key)
-                    }
-                }
-            }
-        }
-        $results.add($keyName, $hash)
-    }
-    else
-    {
-        $platformRestrictions = [Hashtable]::new($Properties)
-        $platformRestrictions.remove('@odata.type')
-        $platformRestrictions.remove('@odata.context')
-        foreach ($key in $platformRestrictions.Keys)
-        {
-            $keyName = $key.Substring(0, 1).toUpper() + $key.substring(1, $key.length - 1)
-            $keyValue = [Hashtable]::new($platformRestrictions.$key)
-            $hash = @{}
-            foreach ($key in $keyValue.Keys)
-            {
-                if ($null -ne $keyValue.$key)
-                {
-                    switch -Wildcard ($keyValue.$key.getType().name)
-                    {
-                        '*[[\]]'
-                        {
-                            if ($keyValue.$key.count -gt 0)
-                            {
-                                $hash.add($key, $keyValue.$key)
-                            }
-                        }
-                        'String'
-                        {
-                            if (-Not [String]::IsNullOrEmpty($keyValue.$key))
-                            {
-                                $hash.add($key, $keyValue.$key)
-                            }
-                        }
-                        Default
-                        {
-                            $hash.add($key, $keyValue.$key)
-                        }
-                    }
-
-                }
-            }
-            $results.add($keyName, $hash)
-        }
-    }
-
-    return $results
 }
 
 Export-ModuleMember -Function *-TargetResource